[Menu + Folder Restructure]
Moved code files into category folders.

Added a Menu class that will be used to create generic menu functionality.

Updated data type of vsyncMode to int16 in both the host application and switch application.

Updated FFMPEGConfigUI to use the new Menu class.

+++ switch-project/source/system/Configuration.cpp
@@ -152,7 +152,7 @@ bool Configuration::SaveFFMPEG(FFMPEG_Config const data)
         return false;
     }
 
-    auto vsync = std::to_string((int)data.vsyncMode);
+    auto vsync = std::to_string(data.vsyncMode);
     if(!ReplaceVariable(newData, VSYNC_TAG, vsync, newData))
     {
         std::cout << "Save error --> couldn't replace " << VSYNC_TAG << std::endl;
